Consolidate WorkflowWaitTrait into WorkflowProcessingTrait
The plugin introduced WorkflowWaitTrait with a TODO noting it should move
to the core WorkflowProcessingTrait. This resolves that TODO by merging
the timeout support and waitForWorkflow() method into the core trait, then
deleting WorkflowWaitTrait and updating Site\CreateCommand to use the core
trait instead.
